pyproject.toml
[project]
name = "playbook-executor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "asyncssh>=2.22.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
]
